apply proper fix and undo imports
