+++ util/Seeder/Scenes/UserLoginCipherScene.cs
@@ -2,7 +2,6 @@
 using Bit.Core.Repositories;
 using Bit.Core.Vault.Entities;
 using Bit.Core.Vault.Repositories;
-using Bit.Seeder.Scenes.LoginCipherScene;
 using Bit.Seeder.Services;
 
 namespace Bit.Seeder.Scenes;
